+++ lib/metrics/stdout.go
@@ -19,12 +19,10 @@ import (
 func init() {
 	Constructors[TypeStdout] = TypeSpec{
 		constructor: NewStdout,
+		Beta:        true,
 		Summary: `
 Prints aggregated metrics as JSON objects to stdout.`,
 		Description: `
-EXPERIMENTAL: This component is considered experimental and is therefore subject
-to change outside of major version releases.
-
 When Benthos shuts down all aggregated metrics are printed. If a
 ` + "`push_interval`" + ` is specified then metrics are also printed
 periodically.`,