+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Improved parser error messages with the `blobl` subcommand.
+- Added flag `file` to the `blobl` subcommand.
+
 ## 3.24.1 - 2020-08-03
 
 ### Fixed